server/requirements.txt
backports-abc==0.4
certifi==2016.9.26
pika==0.10.0
protobuf==3.1.0.post1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
pycrypto